# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/orders":{"get":{"operationId":"OrdersController_getOrders","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VmOrder"}}}}}},"tags":["Orders"]},"post":{"operationId":"OrdersController_addOrder","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VmOrder"}}}}},"tags":["Orders"]}},"/orders/{id}":{"get":{"operationId":"OrdersController_getOrderById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VmOrder"}}}}},"tags":["Orders"]},"put":{"operationId":"OrdersController_updatedOrderById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VmOrder"}}}},"404":{"description":""}},"tags":["Orders"]},"delete":{"operationId":"OrdersController_deleteOrder","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Orders"]}}},"info":{"title":"Orders","description":"Documentation for Orders","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"VmOrder":{"type":"object","properties":{"id":{"type":"string"},"amount":{"type":"number"},"longitude":{"type":"number"},"latitude":{"type":"number"},"payment":{"type":"string"},"clientId":{"type":"string"}},"required":["id","amount","longitude","latitude","payment","clientId"]},"CreateOrderDto":{"type":"object","properties":{"amount":{"type":"number"},"longitude":{"type":"number"},"latitude":{"type":"number"},"payment":{"type":"string"},"clientId":{"type":"string"}},"required":["amount","longitude","latitude","payment","clientId"]},"UpdateOrderDto":{"type":"object","properties":{"id":{"type":"string"},"amount":{"type":"number"},"longitude":{"type":"number"},"latitude":{"type":"number"},"payment":{"type":"string"},"clientId":{"type":"string"}},"required":["id","amount","longitude","latitude","payment","clientId"]}}}}